yun/requirements.txt
requests
pandas
tqdm
kafka-python